Only Educational Purpose
I understand. Google Dorks are a powerful tool that can be used for a variety of purposes, both good and bad. It is important to use them responsibly and ethically.
Here are some tips for using Google Dorks responsibly:

=>Only use Google Dorks to find information that is publicly available. Do not use them to hack into websites or steal personal information.
=>Be respectful of other people's privacy. Do not use Google Dorks to find information about people that they do not want to be public.
=>Use Google Dorks for educational purposes only. Do not use them to harm others or to break the law.

If you are unsure whether or not it is ethical to use Google Dorks for a particular purpose, it is always best to err on the side of caution and not use them.
I am glad that you are using Google Dorks for educational purposes. They can be a valuable tool for learning more about the internet and how it works.